pipeline_ai/requirements.txt
fastapi>=0.111.0
uvicorn>=0.30.1
sqlalchemy>=2.0.30
streamlit>=1.36.0
pandas>=2.2.2
anthropic>=0.30.1
python-dotenv>=1.0.1
